Fixing model sessions and tcp recv in server.py
In current version, the model for hand and fingertip cannot be used simultaneously which leads to an error when one tries to run the code. The error is due to global variables mixing from two models. This commit fixes this error with the usage of two separate tensorflow sessions for the two models.
Also, there can be problems with the call of recv socket function. There are no guarantees that the whole packet will be received, and the code does not take this into account. The commit fixes this by using the socket flag (if it supported on a given platform) or by looping while whole message is not received.